+++ Classes/Exception/AccessDenied.php
@@ -4,7 +4,9 @@
  * This exception can be used to indicate that a user has no access to a page.
  *
  * @author Saskia Metzler <saskia@merlin.owl.de>
+ *
+ * @deprecated will be removed in oelib 4.0
  */
-class Tx_Oelib_Exception_AccessDenied extends Exception
+class Tx_Oelib_Exception_AccessDenied extends \Exception
 {
 }