+++ Classes/Exception/Database.php
@@ -5,6 +5,8 @@
  * error has occurred.
  *
  * @author Oliver Klee <typo3-coding@oliverklee.de>
+ *
+ * @deprecated will be removed in oelib 4.0
  */
 class Tx_Oelib_Exception_Database extends \Exception
 {